cpp: Always define __BUILTIN_VA_ARG_INCR
This is the key used by the system headers to decide that a compiler uses Sun
Studio-style variable arguments, and is what the Devpro cpp(1) has always
defined.

This is necessary when cpp(1) itself needs to preprocess sources which
indirectly include <sys/va_impl.h> which will #error should it be unable to
detect the variable argument list protocol to use.