cpp: Don't try to use non-existent %r in error reporting
Error reporting was using a '%r' that seems, wherever it exists, to
interpret the next argument as part of the format string, and the
argument following that as a va_list of its arguments.

We have no such thing.

Instead, define a verror() to do the line number stuff in the normal
fashion, and have pperror, ppwarn, and yyerror all call it after they've
set up the va_list and done any other manipulation.