Support for spice remote displays
This adds support for Spice remote displays in a similar way
to VNC displays.
Adds: "spice_port", "spice_password" and "spice_opts" options
Also adds: "soundhw" to enable sound
Also adds: "spice" as an info target
If spice_port is set (and > 0) then it will be the remote
display and VNC will not be used.
This also includes fixes to the net.server code to allow
socket closure to do the right thing.