REFACTOR: List snapshots separately from volumes+filesystems
Listing snapshots with full properties takes a significant amount of
time, increasing the time for a 'vmadm list' from less than a second to
almost a minute on some systems. Since we're really only interested in
the snapshot names, issue a separate 'zfs list' for just that and merge
the data afterwards.