joyent/smartos-live#484 Fresh install fails to boot completely
Updated the funtion create_dump() to use the estimated dumpsize.
On servers with 256G of ram or larger dumpadm needs more then 4G of ram
to dump kernel pages. The estimate from dumpadm -e is to small we
double that and use it for the volsize.

We also check to insure that the minimum size for the dumpvol is 1G .

Updated comment as per rmustacc