sysinfo CPU core terminology is confusing
The term "core" is confusing here because it is used to describe
both the physical CPU count ("CPU Physical Cores") and the total
number of vCPUs ("CPU Total Cores"). IMO, "physical cores" should
refer to the actual core count, but that is not something that is
currently enumerated.  That may be a followup change.

I propose renaming the physical CPU count as "CPU Physical
Processors" to better communicate what this number represents.