Support mediatype application/vnd.docker.distribution.manifest.list.v2+json and application/vnd.oci.image.index.v1+json
For docker images support the manifest.list.v2 and oci.image.index.v1 when getting a manifest for a docker image in de registry-v2 client.
When requesting a manifest a manifestList might be returned. In this case select the digest out of the manifestList for the correct platform.

platform needs to be:
  architecture: amd64
  os: linux

if manifest list contains a valid manifest use the digest to get the manifest with image layers.